refactor(header): extract search input to HeaderSearch component
